+++ src/comp.ts
@@ -1,86 +1,91 @@
 import { constructibleStyleSheets } from './style'
-import { empty } from './render'
 
+export const children = Symbol('children')
+export const events = Symbol('events')
 export const created = Symbol('created')
-const updating = Symbol('updating')
 
 export interface CustomElement extends HTMLElement {
   [created]: boolean
-  [updating]?: Promise<void>
+  [children]: { [key: string]: ElemDecl }
+  [events]: { [key: string]: EventDecl }
   observedAttributes?: string[]
-  connectedCallback?(): void
   attributeChangedCallback?(name: string, oldValue: string, newValue: string): void
-  render?(): Element | DocumentFragment
-  updateComplete: Promise<void>
-  requestUpdate(): Promise<void>
+}
+
+declare type Style = HTMLStyleElement | CSSStyleSheet
+
+export interface ElemDecl {
+  id: string
+  sel: string
+}
+
+export interface EventDecl {
+  name: string
+  id: string
+  sel: string
 }
 
 export interface Comp {
-  tag: string,
-  styles?: Array<() => HTMLElement | CSSStyleSheet>
+  tag: string
+  styles?: Array<() => Style>
+  template?: () => HTMLTemplateElement
 }
 
-export function comp({ tag, styles }: Comp): any {
+function appendClone(shadowRoot: ShadowRoot, el: HTMLElement | DocumentFragment): void {
+  shadowRoot.appendChild(el.cloneNode(true))
+}
+
+function pickElements(childEls: { [key: string]: ElemDecl }, shadowRoot: ShadowRoot) {
+  for (const name in childEls) {
+    const { id, sel } = childEls[name]
+    this[name] = id ? shadowRoot.getElementById(id) : shadowRoot.querySelector(sel)
+  }
+}
+
+function startListening(events: { [key: string]: EventDecl }, shadowRoot: ShadowRoot) {
+  for (const method in events) {
+    const { name, id, sel } = events[method]
+    const el = id ? shadowRoot.getElementById(id) : sel ? shadowRoot.querySelector(sel) : shadowRoot
+    el.addEventListener(name, this[method].bind(this))
+  }
+}
+
+export function comp({ tag, styles, template }: Comp): any {
   const styleCount = styles ? styles.length : 0
-  let stylesheets: (CSSStyleSheet | HTMLElement)[]
+  let stylesheets: Style[]
 
   function applyStyle(shadowRoot: ShadowRoot): void {
     if (constructibleStyleSheets)
       (shadowRoot as any).adoptedStyleSheets = stylesheets ||
         (stylesheets = styles.map(style => style() as CSSStyleSheet))
     else
-      for (const style of styles)
-        shadowRoot.appendChild((style() as HTMLElement).cloneNode(true))
-  }
-
-  function clearContent(shadowRoot: ShadowRoot): void {
-    if (constructibleStyleSheets) {
-      empty(shadowRoot)
-    } else {
-      for (let cnt = shadowRoot.childNodes.length - styleCount; cnt; --cnt)
-        shadowRoot.removeChild(shadowRoot.lastChild)
-    }
-  }
-
-  function updateContent(shadowRoot: ShadowRoot, el: Element | DocumentFragment): void {
-    clearContent(shadowRoot)
-    shadowRoot.appendChild(el)
+      for (const style of styles) appendClone(shadowRoot, style() as HTMLStyleElement)
   }
 
   return <T extends new(...args: any[]) => CustomElement>(ctor: T) => {
     class Comp extends ctor {
       constructor(...args: any[]) {
         super()
-        this.attachShadow({ mode: 'open' });
+        this.attachShadow({ mode: 'open' })
+        const shadowRoot = this.shadowRoot
+        if (styleCount) applyStyle(shadowRoot)
+        if (template) {
+          appendClone(shadowRoot, template().content)
+          pickElements.call(this, childEls, shadowRoot)
+        }
+        startListening.call(this, evts, shadowRoot)
         this[created] = true
-      }
-
-      connectedCallback(): void {
-        if (styleCount) applyStyle(this.shadowRoot)
-        this.requestUpdate()
-        if (super.connectedCallback) super.connectedCallback()
       }
 
       attributeChangedCallback(name: string, oldValue: string, newValue: string): void {
         if (oldValue !== newValue) (this as any)[name] = newValue;
         if (super.attributeChangedCallback) super.attributeChangedCallback(name, oldValue, newValue)
       }
-
-      get updateComplete(): Promise<void> {
-        return this[updating] || Promise.resolve()
-      }
-
-      requestUpdate(): Promise<void> {
-        if (!super.render) return
-        let promise = this[updating]
-        if (!promise)
-          promise = this[updating] = Promise.resolve().then(() => {
-            this[updating] = undefined
-            updateContent(this.shadowRoot, super.render())
-          })
-        return promise
-      }
     }
+
+    const { prototype } = Comp
+    const childEls = prototype[children] || (prototype[children] = {})
+    const evts = prototype[events] || (prototype[events] = {})
 
     customElements.define(tag, Comp)