+++ tools/build.js
@@ -1,5 +1,6 @@
-const esbuild = require('@prantlf/esbuild')
+const { startService } = require('@prantlf/esbuild')
 const style = require('./style/esbuild')
+const templ = require('./templ/esbuild')
 
 const task = process.argv[2]
 const watch = task === 'watch'
@@ -44,6 +45,16 @@ if (task === 'test' || watch)
       format: 'cjs'
     },
     {
+      entryPoints: ['test/elem.test.ts'],
+      outfile: 'test/elem.test.js',
+      format: 'cjs'
+    },
+    {
+      entryPoints: ['test/event.test.ts'],
+      outfile: 'test/event.test.js',
+      format: 'cjs'
+    },
+    {
       entryPoints: ['test/prop.test.ts'],
       outfile: 'test/prop.test.js',
       format: 'cjs'
@@ -53,10 +64,15 @@ if (task === 'test' || watch)
       outfile: 'test/example/counter.js',
       format: 'esm',
       bundle: true,
-      plugins: [style({ module: '../..' })]
+      plugins: [style({ module: '../..' }), templ({ module: '../..' })]
     }
   )
 
-Promise
-  .all(builds.map(build => esbuild.build({ ...build, sourcemap: true, watch })))
-  .catch(() => process.exit(1))
+let service
+startService()
+  .then(s => {
+    service = s
+    Promise .all(builds.map(build => service.build({ ...build, sourcemap: true, watch })))
+  })
+  .catch(() => process.exitCode = 1)
+  .finally(() => service && service.stop())