+++ CHANGES.md
@@ -1,5 +1,14 @@
 # Changes
 
+## 0.1.0
+
+* Remove automatic re-rendering on a property change.
+* Remove the `render` class method and asynchronous rendering.
+* Remove the `empty` and `render` helpers.
+* Add `elem` and `event` decorators.
+* Add `templ` - a function memoizing creation of `HTMLTemplateELement`.
+* Add plugins for `eslint` and `rollup` to transform `html` files as functions returning `HTMLTemplateElement`.
+
 ## 0.0.1
 
 Initial release.